requirements.txt
torch
torchvision
# torch==1.8.2+cu111
# torchvision==0.9.2+cu111
# pip3 install torch==1.8.2 torchvision==0.9.2 --extra-index-url https://download.pytorch.org/whl/lts/1.8/cu111
einops==0.4.1
lpips==0.1.4 # eval
opencv_python==4.6.0.66
Pillow
# prettytable==3.3.0 # optional
tensorboardX==2.5.1
timm==0.4.12
tqdm==4.64.0
